% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamapsdata-package.R
\name{am_citation}
\alias{am_citation}
\title{Citation to use when publishing data from this source}
\usage{
am_citation(format = c("text", "md"))
}
\arguments{
\item{format}{one of "text" or "md", default: "text}
}
\value{
citation text and copyright info
}
\description{
You are welcome to include text, numbers and maps from AquaMaps in your
own work for non-commercial use, given that such inserts are clearly
identified as coming from AquaMaps.org, with a backward link to the
respective source.
}
\details{
A researcher planning publication based on these datasets is invited
to contact the AquaMaps team with queries related to specific content.
The AquaMaps team can help with double checking for correctness
prior to drawing conclusions and/or subsequent publication,
and provide further clarification of update frequency,
limitations  and/or interpretation of unlikely results.

Please open an issue on the
\href{https://github.com/raquamaps/aquamapsdata/issues}{GitHub issue tracker} or
contact the team directly by email at Rainer Froese (rfroese (at) geomar.de)
or Kristin Kaschner (Kristin.Kaschner (at) biologie.uni-freiburg.de).
}
\seealso{
Other meta: 
\code{\link{am_meta}},
\code{\link{aquamapsdata-package}}
}
\concept{meta}
