% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{am_csc_from_extent}
\alias{am_csc_from_extent}
\title{Cell identifiers for an extent or bounding box}
\usage{
am_csc_from_extent(x1, x2, y1, y2)
}
\arguments{
\item{x1}{decimal degree WGS84 longitude value for extent (xmin)}

\item{x2}{decimal degree WGS84 longitude value for extent (xmax)}

\item{y1}{decimal degree WGS84 latitude value for extent (ymin)}

\item{y2}{decimal degree WGS84 latitude value for extent (ymax)}
}
\value{
tibble with CsquareCode identifiers
}
\description{
Half degree grid cell identifiers (CsquareCode strings) within an
extent or bounding box.
}
\examples{
\dontrun{
am_csc_from_extent(-5, 5, -7, -5)
}
}
\seealso{
Other spatial: 
\code{\link{am_map_leaflet}()},
\code{\link{am_raster}()},
\code{\link{am_species_in_csc}()},
\code{\link{am_species_per_csc}()},
\code{\link{csquare_from_dd}()},
\code{\link{csquare_to_dd}()},
\code{\link{rasterize_coords}()},
\code{\link{which_cells_in_raster}()}
}
\concept{spatial}
