% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{am_custom_query}
\alias{am_custom_query}
\title{Run a custom SQL query}
\usage{
am_custom_query(sql_query, con, ...)
}
\arguments{
\item{sql_query}{the query}

\item{con}{the connection to use, if missing an sqlite con is used}

\item{...}{other arguments to be passed to the sql() fcn}
}
\description{
Run a custom SQL query
}
\examples{
\dontrun{
am_custom_query("select * from speciesoccursum_r", con = con_am("extdata"))
}
}
\seealso{
Other admin: 
\code{\link{am_db_duckdb}()},
\code{\link{am_db_sqlite}()},
\code{\link{am_use_offline_db}()},
\code{\link{con_am}()},
\code{\link{db_disco}()},
\code{\link{db_minify_path}()},
\code{\link{db_minify}()},
\code{\link{db_sync}()},
\code{\link{default_db}()},
\code{\link{http_ping}()}
}
\concept{admin}
