% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{am_db_sqlite}
\alias{am_db_sqlite}
\title{Location of sqlite3 db file}
\usage{
am_db_sqlite()
}
\value{
character string representing on disk location for db file
}
\description{
Location of sqlite3 db file
}
\seealso{
Other admin: 
\code{\link{am_custom_query}()},
\code{\link{am_db_duckdb}()},
\code{\link{am_use_offline_db}()},
\code{\link{con_am}()},
\code{\link{db_disco}()},
\code{\link{db_minify_path}()},
\code{\link{db_minify}()},
\code{\link{db_sync}()},
\code{\link{default_db}()},
\code{\link{http_ping}()}
}
\concept{admin}
