% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{am_hcaf}
\alias{am_hcaf}
\title{Half degree cell location reference data
This table represents an authority file with reference data for half degree
cell locations.}
\usage{
am_hcaf()
}
\description{
Half degree cell location reference data
This table represents an authority file with reference data for half degree
cell locations.
}
\examples{
\dontrun{
am_hcaf()
}
}
\seealso{
Other general: 
\code{\link{\%>\%}()},
\code{\link{am_hspen}()},
\code{\link{am_search_exact}()},
\code{\link{am_search_fuzzy}()},
\code{\link{download_db}()}
}
\concept{general}
