% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{am_hspen}
\alias{am_hspen}
\title{Environmental envelope preference data for suitable habitats for species
This table provides data on species environmental parameters / preferences}
\usage{
am_hspen()
}
\description{
Environmental envelope preference data for suitable habitats for species
This table provides data on species environmental parameters / preferences
}
\examples{
\dontrun{
keys <- am_search_fuzzy("trevally")$key[1:2]
am_hspen() \%>\% filter(SpeciesID \%in\% keys)
}
}
\seealso{
Other general: 
\code{\link{\%>\%}()},
\code{\link{am_hcaf}()},
\code{\link{am_search_exact}()},
\code{\link{am_search_fuzzy}()},
\code{\link{download_db}()}
}
\concept{general}
