% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{am_map_leaflet}
\alias{am_map_leaflet}
\title{Map for native habitat}
\usage{
am_map_leaflet(
  ras,
  title = "",
  cols = c("#FEB24C", "#FD8D3C", "#FC4E2A", "#E31A1C", "#B10026")
)
}
\arguments{
\item{ras}{a raster object, for example from the am_raster function}

\item{title}{a string to use as legend title, by default blank}

\item{cols}{a vector with three hex colors to use for a numeric color
legend, default: c("#FEB24C", "#FD8D3C", "#FC4E2A", "#E31A1C", "#B10026")}
}
\description{
Shows a leaflet map for the native habitat given a species identifier
}
\examples{
\dontrun{
# native habitat for blue trevally
ras <- am_raster(am_search_fuzzy("blue trevally")$key)
am_map_leaflet(ras, title = "Blue trevally (p)")

# where do both blue and white trevally occur?
ras <- am_raster(am_search_fuzzy("trevally AND (white OR blue)")$key,
  fun = "count")
am_map_leaflet(ras, title = "Blue and white trevally (n)")
}
}
\seealso{
Other spatial: 
\code{\link{am_csc_from_extent}()},
\code{\link{am_raster}()},
\code{\link{am_species_in_csc}()},
\code{\link{am_species_per_csc}()},
\code{\link{csquare_from_dd}()},
\code{\link{csquare_to_dd}()},
\code{\link{rasterize_coords}()},
\code{\link{which_cells_in_raster}()}
}
\concept{spatial}
