% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{am_name_search_exact}
\alias{am_name_search_exact}
\title{Exact search for taxonomic names}
\usage{
am_name_search_exact(
  key = NULL,
  binomial = NULL,
  vernacular = NULL,
  rank_genus = NULL,
  rank_species = NULL,
  rank_family = NULL,
  rank_kingdom = NULL,
  rank_phylum = NULL,
  rank_class = NULL,
  rank_order = NULL
)
}
\arguments{
\item{key}{a database identifier string}

\item{binomial}{a string on the form Genus Species}

\item{vernacular}{a string with a popular name}

\item{rank_genus}{a string with the genus name}

\item{rank_species}{a string with the species name}

\item{rank_family}{a string with the family name}

\item{rank_kingdom}{a string with the kingdom name}

\item{rank_phylum}{a string with the phylum name}

\item{rank_class}{a string with the class name}

\item{rank_order}{a string with the order name}
}
\value{
tibble with matching identifiers...
#' If the function is called with only one argument and NULL as the value, the result will list all unique values that are available
If the function is called with one or more arguments with real values a tibble will be return with keys or identifiers which exactly matches the search
}
\description{
Exact search for taxonomic names
}
\examples{
 am_name_search_exact(vernacular = "Atlantic cod")
 am_name_search_exact(key = "Fis-23638")
}
