% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{am_name_search_fuzzy}
\alias{am_name_search_fuzzy}
\title{Fuzzy search for terms related to taxonomic names}
\usage{
am_name_search_fuzzy(search_term)
}
\arguments{
\item{search_term}{token query, phrase query or NEAR query (see http://www.sqlite.org/fts3.html)}
}
\value{
tibble with matching keys (database identifiers)
}
\description{
Fuzzy search for terms related to taxonomic names
}
\examples{
 am_name_search_fuzzy("cod")
 am_name_search_fuzzy("cod OR shark")
}
