% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{am_raster}
\alias{am_raster}
\title{Raster with native habitat map data for one or several species}
\usage{
am_raster(key, resolution = 0.5, fun = c("count", "last", "first"))
}
\arguments{
\item{key}{a string identifier for the species}

\item{resolution}{the grid resolution in degress, by default 0.5}

\item{fun}{the aggregation fun to use when grid cells are
covered by multiple spatial features, by default their count}
}
\description{
Raster with native habitat map data for one or several species
}
\details{
if the fun parameter is omitted, the probability field will be
used
}
