% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{am_search_fuzzy}
\alias{am_search_fuzzy}
\title{Fuzzy search for terms related to taxonomic names}
\usage{
am_search_fuzzy(search_term)
}
\arguments{
\item{search_term}{token query, phrase query or NEAR query
(see http://www.sqlite.org/fts5.html)}
}
\value{
tibble with matching keys (database identifiers)
}
\description{
Fuzzy search for terms related to taxonomic names
}
\examples{
\dontrun{
am_search_fuzzy("trevally")
am_search_fuzzy("trevally OR animalia")
}
}
