% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{am_species_in_csc}
\alias{am_species_in_csc}
\title{Species occuring in set of grid cell}
\usage{
am_species_in_csc(csc, min_prob = 0)
}
\arguments{
\item{csc}{a vector of strings with CsquareCodes (grid cell identifiers)}

\item{min_prob}{a numeric with the minimum threshold for probable occurrence,
default 0}
}
\description{
Given a set of grid cell identifiers or CsquareCodes, this function
returns a list of the species that occur there
}
\examples{
\dontrun{
# distinct number of species in a specific grid cell
am_species_in_csc("7516:236:1", 0.99)

# other species likely present in the same area as the bluespotted trevally
xt <- am_raster(am_search_fuzzy("bluespotted trevally")$key)@extent
csc <- am_csc_from_extent(xt@xmin, xt@xmax, xt@ymin, xt@ymax)$CsquareCode
am_species_in_csc(csc, min_prob = 0.9)
}
}
\seealso{
Other spatial: 
\code{\link{am_csc_from_extent}()},
\code{\link{am_map_leaflet}()},
\code{\link{am_raster}()},
\code{\link{am_species_per_csc}()},
\code{\link{csquare_from_dd}()},
\code{\link{csquare_to_dd}()},
\code{\link{rasterize_coords}()},
\code{\link{which_cells_in_raster}()}
}
\concept{spatial}
