% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{am_species_per_csc}
\alias{am_species_per_csc}
\title{Total count for number of distinct species across cells, ie "richness"}
\usage{
am_species_per_csc(csc, min_prob = 0, keys = NULL)
}
\arguments{
\item{csc}{a vector of strings with CsquareCodes (grid cell identifiers)}

\item{min_prob}{a numeric with the threshold for probable occurrence,
default 0}

\item{keys}{a vector of string with SpeciesID values}
}
\description{
Global richness map data (i.e. number of different species per cell),
optionally filtered for taxonomy and a preferred probability threshold
}
\examples{
\dontrun{
# distinct number of species in a specific grid cell
am_species_per_csc("7516:236:1", 0.99)

# distinct species count per grid cells in LME 1
csc <- am_hcaf() \%>\% filter(LME == 1) \%>\% collect() \%>\% pull(CsquareCode)
am_species_per_csc(csc, 0.9)

# global richness map data based on taxonomy
keys <- am_search_exact(Family = "Carangidae")$SpeciesID
am_species_per_csc(keys = keys, min_prob = 0.9)
}
}
\seealso{
Other spatial: 
\code{\link{am_csc_from_extent}()},
\code{\link{am_map_leaflet}()},
\code{\link{am_raster}()},
\code{\link{am_species_in_csc}()},
\code{\link{csquare_from_dd}()},
\code{\link{csquare_to_dd}()},
\code{\link{rasterize_coords}()},
\code{\link{which_cells_in_raster}()}
}
\concept{spatial}
