% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{am_sql_query}
\alias{am_sql_query}
\title{Run a custom SQL query}
\usage{
am_sql_query(sql_query, ...)
}
\arguments{
\item{sql_query}{the query}

\item{...}{other arguments to be passed to the sql() fcn}
}
\description{
Run a custom SQL query
}
