% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamapsdata-package.R
\name{aquamapsdata-package}
\alias{aquamapsdata-package}
\alias{aquamapsdata}
\title{aquamapsdata}
\description{
This package provides data access to curated data from aquamaps.org.
}
\details{
A local SQLite db is downloaded from
https://archive.org/download/aquamapsdb

This database is then exposed in this package for querying (using dplyr)
}
\seealso{
Other meta: 
\code{\link{am_citation}()},
\code{\link{am_meta}}
}
\concept{meta}
\keyword{package}
