% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aquamapsdata_sqlite}
\alias{aquamapsdata_sqlite}
\title{SQLite database connection to the aquamaps data using dplyr}
\usage{
aquamapsdata_sqlite()
}
\description{
SQLite database connection to the aquamaps data using dplyr
}
