% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{con_am}
\alias{con_am}
\title{Connection to aquamapsdata}
\usage{
con_am(source_type = c("sqlite", "mysql", "duckdb"))
}
\arguments{
\item{source_type}{one of "duckdb", "sqlite" or "mysql" with "sqlite" being default}
}
\value{
database connection
}
\description{
This function returns a db connection to one of two possible pre-configured
data sources containing aquamaps data
}
