% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_disco}
\alias{db_disco}
\title{Disconnect the default database connection}
\usage{
db_disco(env = db_cache)
}
\arguments{
\item{env}{the environment holding the connection, by default db_cache}
}
\description{
Disconnect the default database connection
}
