% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_minify}
\alias{db_minify}
\title{Minify a larger aquamaps dataset by slicing}
\usage{
db_minify(key, slice_file)
}
\arguments{
\item{key}{the species identifier to keep}

\item{slice_file}{the file path for storing the resulting SQLite3 db}
}
\description{
Creates a minimized version of a larger AquaMaps database,
by filtering all tables to contain a single identifier.
}
