% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_minify_path}
\alias{db_minify_path}
\title{Location of minified bundled sqlite db}
\usage{
db_minify_path()
}
\description{
A minified sqlite db is bundled in the package for technical reasons.
}
\seealso{
Other admin: 
\code{\link{am_custom_query}()},
\code{\link{am_db_duckdb}()},
\code{\link{am_db_sqlite}()},
\code{\link{am_use_offline_db}()},
\code{\link{con_am}()},
\code{\link{db_disco}()},
\code{\link{db_minify}()},
\code{\link{db_sync}()},
\code{\link{default_db}()},
\code{\link{http_ping}()}
}
\concept{admin}
