% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{default_db}
\alias{default_db}
\title{Set or switch the default database used}
\usage{
default_db(source = Sys.getenv("AM_DB_SOURCE"))
}
\arguments{
\item{source}{string, one of "sqlite", "duckdb", "mysql" or "extdata"}
}
\description{
The database connection used when a specific connection is not provided
can be set with this function. It can also be switched.
}
\details{
the "extdata" source refers to a minified < 1MB sqlite3 db
with a very small subset of the full data, which is bundled into the
package and which allows tests and vignettes to run in the package without
having the full dataset installed locally (during for staged installation).
}
