% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_db}
\alias{download_db}
\title{Perform downloading of compressed SQLite db, storing it locally}
\usage{
download_db(force = FALSE, passphrase = NULL)
}
\arguments{
\item{force}{boolean to indicate whether to overwrite an existing db}

\item{passphrase}{passphrase if using encrypted db, default: NULL}
}
\description{
Perform downloading of compressed SQLite db, storing it locally
}
\examples{
\dontrun{
download_db()
}
}
\seealso{
Other general: 
\code{\link{\%>\%}()},
\code{\link{am_hcaf}()},
\code{\link{am_hspen}()},
\code{\link{am_search_exact}()},
\code{\link{am_search_fuzzy}()}
}
\concept{general}
