% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{http_ping}
\alias{http_ping}
\title{Ping a web site}
\usage{
http_ping(url = "http://aquamaps.org", timeout = 10)
}
\arguments{
\item{url}{a web site to "web ping" to see that it is available}
}
\value{
TRUE if web site is up, FALSE otherwise
}
\description{
Ping a web site
}
\examples{
http_ping("http://aquamaps.org")
}
