% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{pipe - convert nested functions into sequential statements}
\description{
This package uses the pipe function, \verb{\\\%>\\\%} to turn
function composition into a series of imperative statements.
}
