% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rasterize_coords}
\alias{rasterize_coords}
\title{Converts location coordinate data into raster data}
\usage{
rasterize_coords(lon, lat, fun = "count")
}
\arguments{
\item{lon}{decimal longitude for coordinates}

\item{lat}{decimal latitude for coordinates}

\item{fun}{aggregation method when several points occur in one cell}
}
\value{
a raster grid with individual cell values representing
the count of occurrences within the grid cells using WGS84
}
\description{
Converts location coordinate data into raster data
}
\seealso{
Other spatial: 
\code{\link{am_csc_from_extent}()},
\code{\link{am_map_leaflet}()},
\code{\link{am_raster}()},
\code{\link{am_species_in_csc}()},
\code{\link{am_species_per_csc}()},
\code{\link{csquare_from_dd}()},
\code{\link{csquare_to_dd}()},
\code{\link{which_cells_in_raster}()}
}
\concept{spatial}
