% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{remote_update}
\alias{remote_update}
\title{Perform remote update by downloading zip with CSV and migrating files to a SQLite db stored locally}
\usage{

  remote_update(AM_SRC = "https://ftp.geomar.de/users/cgarilao/AquaMaps_ver0816.zip",
  AM_DB = paste0(system.file(package = "aquamapsdata"), "/extdata/am.db"),
  truncate = TRUE)
}
\arguments{
\item{AM_SRC}{URL for zip of CSV files with raw data}

\item{AM_DB}{location of SQLite db locally}

\item{truncate}{boolean indicating whether to truncate CSVs
when converting to SQLite (used for debug and performance purposes)}
}
\description{
Perform remote update by downloading zip with CSV and migrating files to a SQLite db stored locally
}
