% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{which_cells_in_raster}
\alias{which_cells_in_raster}
\title{Raster grid cell identifiers (LOICZID) with presence}
\usage{
which_cells_in_raster(r, lower_limit = 0)
}
\arguments{
\item{r}{raster (needs to be half degree cell raster WGS84 720x360)}

\item{lower_limit}{raster cell values higher than this parameter
is interpreted as presence}
}
\value{
a vector of raster grid cell identifiers (loiczids)
}
\description{
Raster grid cell identifiers (LOICZID) with presence
}
\seealso{
Other spatial: 
\code{\link{am_csc_from_extent}()},
\code{\link{am_map_leaflet}()},
\code{\link{am_raster}()},
\code{\link{am_species_in_csc}()},
\code{\link{am_species_per_csc}()},
\code{\link{csquare_from_dd}()},
\code{\link{csquare_to_dd}()},
\code{\link{rasterize_coords}()}
}
\concept{spatial}
