% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{am_custom_query}
\alias{am_custom_query}
\title{Run a custom SQL query}
\usage{
am_custom_query(sql_query, con, ...)
}
\arguments{
\item{sql_query}{the query}

\item{con}{the connection to use, if missing an sqlite con is used}

\item{...}{other arguments to be passed to the sql() fcn}
}
\description{
Run a custom SQL query
}
