% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{am_map_leaflet}
\alias{am_map_leaflet}
\title{Map for native habitat}
\usage{
am_map_leaflet(ras, title = "", cols = c("#FFFFCC", "#0C2C84", "#41B6C4"))
}
\arguments{
\item{ras}{a raster object, for example from the am_raster function}

\item{title}{a string to use as legend title, by default blank}

\item{cols}{a vector with three hex colors to use for a numeric color
legend, default:}
}
\description{
Shows a leaflet map for the native habitat given a species identifier
}
