% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{am_raster}
\alias{am_raster}
\title{Raster with native habitat map data for one or several species}
\usage{
am_raster(key, resolution = 0.5, fun = c("count", "last", "first"))
}
\arguments{
\item{key}{a string identifier for the species}

\item{resolution}{the grid resolution in degrees, by default 0.5}

\item{fun}{the aggregation fun to use when grid cells are
covered by multiple spatial features, by default their count}
}
\description{
Raster with native habitat map data for one or several species
}
\details{
if the fun parameter is omitted, the probability field will be
used
}
\examples{
\dontrun{
ras <- am_raster(am_search_fuzzy("trevally")$key)
}
}
\seealso{
Other spatial: 
\code{\link{am_csc_from_extent}()},
\code{\link{am_map_leaflet}()},
\code{\link{am_species_in_csc}()},
\code{\link{am_species_per_csc}()},
\code{\link{csquare_from_dd}()},
\code{\link{csquare_to_dd}()},
\code{\link{rasterize_coords}()},
\code{\link{which_cells_in_raster}()}
}
\concept{spatial}
