% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamapsdata-package.R
\name{aquamapsdata-package}
\alias{aquamapsdata-package}
\alias{aquamapsdata}
\title{aquamapsdata}
\description{
This package provides data access to curated data from aquamaps.org.
}
\details{
A local SQLite db is created based on raw data files from https://ftp.geomar.de/users/cgarilao/AquaMaps_ver0816.zip

This database is then exposed in this package for querying (using dplyr)
}
\keyword{package}
