% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.R
\name{csquare_from_dd}
\alias{csquare_from_dd}
\title{csquare code conversion from decimal degrees}
\usage{
csquare_from_dd(lat, lon, resolution = 1e-04)
}
\arguments{
\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}

\item{resolution}{the resolution in decimal degress
(one of 0.1, 0.01, 0.001, 0.0001 etc)}
}
\value{
a character string with the code
}
\description{
The csquare code is explained here:
http://www.cmar.csiro.au/csquares/csq-faq.htm
}
\examples{
csquare_from_dd(50.93578, -114.01435, resolution = 0.00001)
}
\seealso{
Other spatial: 
\code{\link{am_csc_from_extent}()},
\code{\link{am_map_leaflet}()},
\code{\link{am_raster}()},
\code{\link{am_species_in_csc}()},
\code{\link{am_species_per_csc}()},
\code{\link{csquare_to_dd}()},
\code{\link{rasterize_coords}()},
\code{\link{which_cells_in_raster}()}
}
\concept{spatial}
