% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csquare.R
\name{csquare_to_dd}
\alias{csquare_to_dd}
\title{csquare code conversion to decimal degrees}
\usage{
csquare_to_dd(code)
}
\arguments{
\item{code}{csquare code}
}
\value{
a list with decimal degrees lat and lon and resolution
}
\description{
Given a csquare code encoded coordinate this function converts it
into decimal degrees
}
\details{
The csquare code is explained here:
http://www.cmar.csiro.au/csquares/csq-faq.htm
}
\examples{
\dontrun{
if(interactive()){
 csquare_to_dd("7307:487:380:383")
 }
}
}
\seealso{
Other csquare: 
\code{\link{csquare_from_dd}()}
}
\concept{csquare}
