% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_minify}
\alias{db_minify}
\title{Minify a larger aquamaps dataset by slicing}
\usage{
db_minify(key, slice_file)
}
\arguments{
\item{key}{the species identifier to keep}

\item{slice_file}{the file path for storing the resulting SQLite3 db}
}
\description{
Creates a minimized version of a larger AquaMaps database,
by filtering all tables to contain a single identifier.
}
\seealso{
Other admin: 
\code{\link{am_custom_query}()},
\code{\link{am_db_duckdb}()},
\code{\link{am_db_sqlite}()},
\code{\link{am_use_offline_db}()},
\code{\link{con_am}()},
\code{\link{db_disco}()},
\code{\link{db_minify_path}()},
\code{\link{db_sync}()},
\code{\link{default_db}()},
\code{\link{http_ping}()}
}
\concept{admin}
