% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{download_db}
\alias{download_db}
\title{Perform downloading of compressed SQLite db, storing it locally}
\usage{
download_db(force = FALSE)
}
\arguments{
\item{force}{boolean to indicate whether to overwrite an existing db}
}
\description{
Perform downloading of compressed SQLite db, storing it locally
}
