% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{http_ping}
\alias{http_ping}
\title{Ping a web site}
\usage{
http_ping(url = "http://aquamaps.org", timeout = 10)
}
\arguments{
\item{url}{a web site to "web ping" to see that it is available}

\item{timeout}{default time to wait in seconds before returning FALSE}
}
\value{
TRUE if web site is up, FALSE otherwise
}
\description{
Ping a web site
}
\examples{
\dontrun{
 http_ping("http://aquamaps.org")
}

}
\seealso{
Other admin: 
\code{\link{am_custom_query}()},
\code{\link{am_db_duckdb}()},
\code{\link{am_db_sqlite}()},
\code{\link{am_use_offline_db}()},
\code{\link{con_am}()},
\code{\link{db_disco}()},
\code{\link{db_minify_path}()},
\code{\link{db_minify}()},
\code{\link{db_sync}()},
\code{\link{default_db}()}
}
\concept{admin}
